refactor(#414): extract TrackedFileCard from ChatConsole
- Move TrackedFile type + OFFICE_FILE_RE_LEGACY → TrackedFileCard.tsx
- Extract TrackedFileCard component (-116 lines from ChatConsole)
- Total removed from ChatConsole: ~250 lines (Avatars + MarkdownContent + DiffView + TrackedFileCard)
